Unify "Capability"
