+++ src/Core/Addon/Capability/ICanLoadAddons.php
@@ -19,25 +19,19 @@
  *
  */
 
-namespace Friendica\Core\Logger\Capabilities;
+namespace Friendica\Core\Addon\Capability;
 
 /**
- * An enum class for the Log channels
+ * Interface for loading Addons specific content
  */
-interface LogChannel
+interface ICanLoadAddons
 {
-	/** @var string channel for the auth_ejabbered script */
-	public const AUTH_JABBERED = 'auth_ejabberd';
-	/** @var string Default channel in case it isn't set explicit */
-	public const DEFAULT = self::APP;
-	/** @var string channel for console execution */
-	public const CONSOLE = 'console';
-	/** @var string channel for developer focused logging */
-	public const DEV = 'dev';
-	/** @var string channel for daemon executions */
-	public const DAEMON = 'daemon';
-	/** @var string channel for worker execution */
-	public const WORKER = 'worker';
-	/** @var string channel for frontend app executions */
-	public const APP = 'app';
+	/**
+	 * Returns a merged config array of all active addons for a given config-name
+	 *
+	 * @param string $configName The config-name (config-file at the static directory, like 'hooks' => '{addon}/static/hooks.config.php)
+	 *
+	 * @return array the merged array
+	 */
+	public function getActiveAddonConfig(string $configName): array;
 }
