+++ tests/src/Core/KeyValueStorage/KeyValueStorageTest.php
@@ -21,7 +21,7 @@
 
 namespace Friendica\Test\src\Core\KeyValueStorage;
 
-use Friendica\Core\KeyValueStorage\Capabilities\IManageKeyValuePairs;
+use Friendica\Core\KeyValueStorage\Capability\IManageKeyValuePairs;
 use Friendica\Test\MockedTest;
 
 abstract class KeyValueStorageTest extends MockedTest
